# apps/desktop/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eigenheim engine",
    "version": "0.1.0"
  },
  "paths": {
...
    "/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_events_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response Events Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/syncs": {
      "get": {
        "summary": "Syncs",
        "operationId": "syncs_syncs_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SyncOut"
                  },
                  "title": "Response Syncs Syncs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/hypotheses": {
      "get": {
        "summary": "List Hypotheses",
        "description": "List hypotheses with optional filters.\n\nstatus: proposed | testing | confirmed | rejected (omit = all).\nlogic_id: filter to hypotheses linked to a specific metric (omit = all).",
        "operationId": "list_hypotheses_hypotheses_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "logic_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Logic Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HypothesisOut"
                  },
                  "title": "Response List Hypotheses Hypotheses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Hypothesis",
        "description": "Create a hypothesis (human path; source='user', status='proposed').",
        "operationId": "create_hypothesis_hypotheses_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HypothesisIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HypothesisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hypotheses/{hyp_id}": {
      "get": {
        "summary": "Get Hypothesis",
        "description": "Get one hypothesis by id.",
        "operationId": "get_hypothesis_hypotheses__hyp_id__get",
        "parameters": [
          {
            "name": "hyp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hyp Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HypothesisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Hypothesis",
        "description": "Delete a hypothesis.",
        "operationId": "delete_hypothesis_hypotheses__hyp_id__delete",
        "parameters": [
          {
            "name": "hyp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hyp Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hypotheses/{hyp_id}/status": {
      "patch": {
        "summary": "Update Hypothesis Status",
        "description": "Advance the status of a hypothesis (human-only gate).",
        "operationId": "update_hypothesis_status_hypotheses__hyp_id__status_patch",
        "parameters": [
          {
            "name": "hyp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hyp Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HypothesisStatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HypothesisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/graph/build": {
      "post": {
        "summary": "Build Graph",
        "description": "POST /graph/build \u2014 run graphify over a directory and return a run_id.",
        "operationId": "build_graph_graph_build_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphBuildRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphBuildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graph/{run_id}/json": {
      "get": {
        "summary": "Get Graph Json",
        "description": "GET /graph/{run_id}/json \u2014 serve the capped, metric-enriched graph JSON.",
        "operationId": "get_graph_json_graph__run_id__json_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DeletedOut": {
        "properties": {
          "deleted": {
            "type": "string",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "DeletedOut"
      },
...
      "EventOut": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "origin",
          "source",
          "description"
        ],
        "title": "EventOut"
      },
...
      "GraphBuildRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "infer": {
            "type": "boolean",
            "title": "Infer",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "GraphBuildRequest"
      },
      "GraphBuildResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "started_at"
        ],
        "title": "GraphBuildResponse"
      },
      "GraphEdge": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "relation": {
            "type": "string",
            "title": "Relation"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target",
          "relation",
          "confidence"
        ],
        "title": "GraphEdge"
      },
      "GraphNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "source_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source File"
          },
          "community": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Community"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "trace_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Ref"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "GraphNode"
      },
      "GraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "built_at": {
            "type": "string",
            "title": "Built At"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges",
          "run_id",
          "built_at"
        ],
        "title": "GraphResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HypothesisIn": {
        "properties": {
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "logic_id": {
            "type": "string",
            "title": "Logic Id",
            "default": ""
          },
          "evidence": {
            "type": "string",
            "title": "Evidence",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "statement"
        ],
        "title": "HypothesisIn"
      },
      "HypothesisOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "logic_id": {
            "type": "string",
            "title": "Logic Id"
          },
          "evidence": {
            "type": "string",
            "title": "Evidence"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "statement",
          "logic_id",
          "evidence",
          "status",
          "source",
          "created_at"
        ],
        "title": "HypothesisOut"
      },
      "HypothesisStatusIn": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HypothesisStatusIn"
      },
...
      "SyncHistoryItem": {
        "properties": {
          "at": {
            "type": "string",
            "title": "At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "at",
          "status",
          "detail"
        ],
        "title": "SyncHistoryItem"
      },
      "SyncOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          },
          "nextRun": {
            "type": "string",
            "title": "Nextrun"
          },
          "lastStatus": {
            "type": "string",
            "title": "Laststatus"
          },
          "lastRun": {
            "type": "string",
            "title": "Lastrun"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/SyncHistoryItem"
            },
            "type": "array",
            "title": "History",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "target",
          "frequency",
          "nextRun",
          "lastStatus",
          "lastRun"
        ],
        "title": "SyncOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}